# internal/docs/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Sarbon API
  description: |
    Закрытый backend — доступ только для доверенных клиентов (frontend или mobile).
    На каждый запрос обязательны заголовки (токен только в header X-Client-Token):
    - X-Client-Type (frontend | mobile)
    - X-Platform (web | ios | android)
    - X-Client-Token — токен приложения (обязателен для всех API)
    - Accept-Language (ru | en | uz | tr | zh)
    Authorization: Bearer &lt;JWT&gt; — только для API, где в описании указано «требуется JWT».
  version: 1.0.0
  x-response-format: |
    Все ответы API (JSON) имеют единый формат: status (число — HTTP-код), message (строка — "success" или текст ошибки), data (тело ответа или null при ошибке).
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /status-codes:
    get:
      summary: Справочник кодов статуса
      description: Возвращает список всех кодов статуса API и соответствующих сообщений в одном запросе. Единый формат ответа (status, message, data); в data — массив объектов { code, message }.
      operationId: statusCodes
      tags: [system]
      security:
        - ClientToken: []
      parameters:
        - $ref: '#/components/parameters/XClientType'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StatusCodeItem'
...
  /reference/user-categories:
    get:
      summary: Список категорий пользователей
      description: Справочник категорий (super_admin, admin, carrier, top_dispatcher, dispatcher, driver, vehicle, manager). Доступ по X-Client-Token (frontend или mobile).
      operationId: userCategories
      tags: [Reference]
      security:
        - ClientToken: []
      parameters:
        - $ref: '#/components/parameters/XClientType'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserCategory'
        '500':
          description: Internal error
...
components:
  schemas:
...
    StatusCodeItem:
      type: object
      description: Элемент справочника GET /status-codes.
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
...
    UserCategory:
      type: object
      properties:
        id:
          type: integer
          example: 1
        code:
          type: string
          example: admin
        name:
          type: string
          example: Admin
        sort_order:
          type: integer
          example: 2
  securitySchemes:
    ClientToken:
      type: apiKey
      in: header
      name: X-Client-Token
      description: Токен приложения — обязателен для каждого запроса. Укажи FRONTEND_CLIENT_TOKEN или MOBILE_CLIENT_TOKEN.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT пользователя — только для API с пометкой «требуется JWT».
  parameters:
    XClientType:
      name: X-Client-Type
      in: header
      required: true
      schema:
        type: string
        enum: [frontend, mobile]
      description: Тип клиента
    XPlatform:
      name: X-Platform
      in: header
      required: true
      schema:
        type: string
        enum: [web, ios, android]
      description: Платформа
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: true
      schema:
        type: string
        enum: [ru, en, uz, tr, zh]
      description: Язык (только из заголовка)